tests: check for first memory report
Getting the memory report shows that the main loop is running.

To be able to check it efficiently, print the first memory report
immediately instead of waiting one second.

Also, drop the check for "Could not set priority". This message
is not printed on Travis for some reason, and it would not be
printed when running the tests as root.