kill: when we have swap space, give the kernel extra time
With swap enabled, the kernel seems to need more than 100ms to free the memory
of the killed process. This means that earlyoom would immediately kill another
process. Sleep a little extra to give the kernel time to free the memory.
(Yes, this will sleep even if the kill has failed. Does no harm and keeps the
code simple.)