kill_wait: escalate to SIGKILL *within* the wait loop
Also, move the usleep to *after* the memory check for
faster reaction time.

We used to return from the kill_wait loop when we drop
below the SIGKILL limits and would immediately be called
again. This means a new victim would be selected, which
may be the wrong one because the original victim is in
the process of being cleaned up, did not release the memory
yet, but does not have a VmRSS value in /proc anymore.