add --prefer-only option to only kill processes matching --prefer
While --avoid exists to list processes that should not be killed, it may
be desirable to have earlyoom simply ignore any processes that do not
match the regex specified in --prefer. It should be noted that the Linux
kernel may still opt to kill other processes if earlyoom did not free up
enough memory to prevent an oom.

Signed-off-by: Murat Boyar <murat.boyar@siemens.com>
Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>