Makefile: factor in PREFIX to install directories
Sometimes we want to set PREFIX and leave the installation directories relative
to the PREFIX.

Other times we want to set some of these installation directories to absolute
values.

Sometimes we want both.

This commit patch achieves this with a reasonably small effort, by not
propagating PREFIX around the Makefile.