Export EARLYOOM_ENVIRON of killed processes along with related info
With this change, along with info such as cmdline, PID, UID and so on,
extra information (its environment variables) will be available to
allow to identify the process being killed due to memory overuse.

This is not captured by the existing methods, and it can be useful to
identify processes which are invoked in the same generic way and whose
runtime behavior or the data that it processes changes depending on
the environment at the time of launch, or that simply by looking at
the content of the variables in its environment gives extra info about
a single process from others using the same binary.

This can be used in combination with -N to run a script with info
about the process just killed, and use the information on the
desired environment variables to log/notify about the problem.